refactor: clean up generationsRouter
